Support for `pop()` member of array type

+++ test/libsolidity/semanticTests/viaYul/array_storage_push_empty_length_address.sol
@@ -1,6 +1,6 @@
 contract C {
     address[] addressArray;
-    function set_get_length(uint256 len) public returns (uint256)
+    function set_get_length(uint len) public returns (uint)
     {
         while(addressArray.length < len)
             addressArray.push();
@@ -13,11 +13,11 @@ contract C {
 // compileViaYul: also
 // EVMVersion: >=petersburg
 // ----
-// set_get_length(uint256): 0 -> 0
-// set_get_length(uint256): 1 -> 1
-// set_get_length(uint256): 10 -> 10
-// set_get_length(uint256): 20 -> 20
-// set_get_length(uint256): 0 -> 0
-// set_get_length(uint256): 0xFF -> 0xFF
-// set_get_length(uint256): 0xFFF -> 0xFFF
-// set_get_length(uint256): 0xFFFF -> FAILURE # Out-of-gas #
+// set_get_length(uint): 0 -> 0
+// set_get_length(uint): 1 -> 1
+// set_get_length(uint): 10 -> 10
+// set_get_length(uint): 20 -> 20
+// set_get_length(uint): 0 -> 0
+// set_get_length(uint): 0x00FF -> 0x00FF
+// set_get_length(uint): 0x0FFF -> 0x0FFF
+// set_get_length(uint): 0x00FFFF -> FAILURE, 5 # Out-of-gas #
